hf-gemini-space/requirements.txt
google-generativeai>=0.8.0
gradio>=5.0.0
